# Conflicts:
#	.idea/deploymentTargetSelector.xml
#	app/src/main/java/com/example/optivote/ViewModel/VoteRecordViewModel.kt
#	app/src/main/java/com/example/optivote/repository/VoteRecordRepository.kt
#	app/src/main/java/com/example/optivote/repository/VoteRecordRepositoryImp.kt

+++ app/src/main/java/com/example/optivote/model/User.kt
@@ -1,6 +1,6 @@
 package com.example.optivote.model
 
-class User (val id : Int?,
+class User (val id : Long?,
             val name:String?,
             val email:String?,
             val phone:String?,
